pyproject.toml
[project]
name = "deployment-simulator"
version = "0.1.0"
description = "Multi-agent orchestration system for simulating enterprise AI deployments"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic[bedrock]>=0.40.0",
    "boto3>=1.35.0",
    "chromadb>=0.5.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
frontend = [
    "streamlit>=1.40",
    "plotly>=5.24",
]

[project.scripts]
simulate = "src.main:main"
